move WriteToUs to components folder
