+++ src/components/WalletTopup/WriteToUs/index.js
@@ -1,0 +1,1 @@
+export { default } from './WriteToUs.vue'
